% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_comp.R
\name{artist_comp}
\alias{artist_comp}
\title{Compares given artist's popularity on Spotify with their popularity on YouTube}
\usage{
artist_comp(url)
}
\arguments{
\item{url}{Link to Spotify artist profile, including `"https://"` at beginning.}
}
\value{
This function returns a list containing:
*A list providing the given artist's overall popularity score on Spotify,
total follower count on Spotify, and subscriber count on YouTube.
*A data frame where each row represents one of
the given artist's top songs on Spotify. There are columns for song name,
Spotify popularity score, YouTube view count, match confidence score (i.e.
confidence in the match between the Spotify song and YouTube video), Spotify
 popularity rank within the top songs, and YouTube view count rank within
 the top songs.
}
\description{
Compares given artist's popularity on Spotify with their popularity on YouTube
}
