% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_creds.R
\name{auth_creds}
\alias{auth_creds}
\title{Saves YouTube and Spotify API credentials}
\usage{
auth_creds(spotify_id, spotify_secret, yt_client, yt_secret)
}
\arguments{
\item{spotify_id}{Your Spotify API client ID.}

\item{spotify_secret}{Your Spotify API client secret.}

\item{yt_key}{Your YouTube API key.}
}
\description{
This function must be run before the other SpoTubeR functions can be used.
You will need to first sign up for a Spotify API account
[here](https://developer.spotify.com/documentation/web-api) (requires a
Spotify Premium subscription) and obtain an API key for the YouTube API
[here](https://developers.google.com/youtube/v3/). After running this function,
 your credentials will be saved to your R environment, and SpoTubeR's functions
  will automatically retrieve them, even across R sessions. (Currently, the
  user may need to restart R in order for the credentials to become retrievable
   by other functions.)
}
\examples{
\dontrun{auth_creds(spotify_id = "YOUR ID HERE", spotify_secret = "YOUR SECRET HERE", yt_key = "YOUR SECRET HERE")}
}
