% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{billboard100_apr26}
\alias{billboard100_apr26}
\title{Billboard Hot 100 dataset}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{rank}{song rank}
  \item{track_name}{song name}
  \item{artist_name}{name of song artist}
  ...
}
}
\source{
[Billboard Hot 100](https://www.billboard.com/charts/hot-100/)
}
\usage{
billboard100_apr26
}
\description{
A dataset with the Billboard Hot 100 songs as of April 11th, 2026.
}
\keyword{datasets}
