% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_yt_song.R
\name{match_yt_song}
\alias{match_yt_song}
\title{Matches YouTube music video to corresponding Spotify song}
\usage{
match_yt_song(url)
}
\arguments{
\item{url}{Link to YouTube music video, in the format beginning with
\code{"https://www.youtube.com/watch?v="}.}
}
\value{
This function returns a list containing the Spotify popularity score,
corresponding YouTube video view count, and match confidence score (i.e. the
confidence that the correct YouTube video was matched to the given Spotify
song).
}
\description{
Takes a YouTube music video URL and returns a list with data comparing the song's
popularity between YouTube and Spotify
}
\details{
Before running this function, make sure you've set your API
credentials using \code{\link[=auth_creds]{auth_creds()}}.
}
\examples{
# remember you must set up your API credentials with [auth_creds()] before you can run examples
# currently produces error when over YouTube API quota limit, so designated as don't run for now
# using Ordinary by Alex Warren
\dontrun{
match_yt_song("https://www.youtube.com/watch?v=u2ah9tWTkmk")}
}
