% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_artist_comp.R
\name{spotify_artist_comp}
\alias{spotify_artist_comp}
\title{Compares given artist's popularity on Spotify with their popularity on YouTube}
\usage{
spotify_artist_comp(url)
}
\arguments{
\item{url}{Link to Spotify artist profile, including \code{"https://"} at beginning.}
}
\value{
This function returns a list containing:
\itemize{
\item A list providing the given artist's overall popularity score on Spotify,
total follower count on Spotify, and subscriber count on YouTube. (If YouTube
subscriber count is hidden, the value will appear as \code{NA}).
\item A data frame where each row represents one of
the given artist's top songs on Spotify. There are columns for song name,
Spotify popularity score, YouTube view count, match confidence score (i.e.
confidence in the match between the Spotify song and YouTube video), Spotify
popularity rank within the top songs, and YouTube view count rank within
the top songs.
}
}
\description{
This function takes a Spotify artist profile URL as its argument and returns
a list with data comparing the artist's popularity between Spotify and YouTube.
}
\details{
Before running this function, make sure you've set your API
credentials using \code{\link[=auth_creds]{auth_creds()}}.
}
\examples{
# you must set up your API credentials with the auth_creds function first
# using Noah Kahan
spotify_artist_comp("https://open.spotify.com/artist/2RQXRUsr4IW1f3mKyKsy4B?
si=69yU_685T96XI2mWtOGfLg")
}
