% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_artist_comp.R
\name{yt_artist_comp}
\alias{yt_artist_comp}
\title{Compares given artist's popularity on YouTube with their popularity on Spotify}
\usage{
yt_artist_comp(url)
}
\arguments{
\item{url}{Link to artist's YouTube channel, in the format beginning with
\code{"https://www.youtube.com/channel/"}.}
}
\value{
This function returns a list containing:
\itemize{
\item A list providing the given artist's overall popularity score on Spotify,
total follower count on Spotify, and subscriber count on YouTube. (If YouTube
subscriber count is hidden, the value will appear as \code{NA}).
\item A data frame where each row represents one of
the given artist's top songs on Spotify. There are columns for song name,
Spotify popularity score, YouTube view count, match confidence score (i.e.
confidence in the match between the Spotify song and YouTube video), Spotify
popularity rank within the top songs, and YouTube view count rank within
the top songs.
}
}
\description{
This function takes an artist's YouTube channel URL as its argument and returns
a list with data comparing the artist's popularity between YouTube and Spotify
}
\details{
Before running this function, make sure you've set your API
credentials using \code{\link[=auth_creds]{auth_creds()}}.
}
\examples{
# remember you must set up your API credentials with the auth_creds function before you can run examples
# using Noah Kahan
yt_artist_comp("https://www.youtube.com/channel/UCXY5pi3MbsaP1WEgClmglsA")
# using Kacey Musgraves
yt_artist_comp("https://www.youtube.com/channel/UCAxUOuaNs8e63_MRqhTdmaQ")
}
