% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_creds.R
\name{auth_creds}
\alias{auth_creds}
\title{Saves YouTube and Spotify API credentials}
\usage{
auth_creds(spotify_id, spotify_secret, yt_key)
}
\arguments{
\item{spotify_id}{Your Spotify API client ID (as character type).}

\item{spotify_secret}{Your Spotify API client secret (as character type).}

\item{yt_key}{Your YouTube API key (as character type).}
}
\value{
This function does not return anything. It does provide a message
informing the user that their credentials have been saved.
}
\description{
This function must be run before the other SpoTubeR functions can be used.
}
\details{
You will need to first sign up for a Spotify API account
\href{https://developer.spotify.com/documentation/web-api}{here} (requires a
Spotify Premium subscription) and obtain an API key for the YouTube API
\href{https://developers.google.com/youtube/v3/}{here}. After running this function,
your credentials will be saved to your R environment, and SpoTubeR's functions
will automatically retrieve them, even across R sessions. (Currently, the
user may need to restart R in order for the credentials to become retrievable
by other functions.)
}
\examples{
\dontrun{auth_creds(spotify_id = "YOUR ID HERE",
spotify_secret = "YOUR SECRET HERE",
yt_key = "YOUR SECRET HERE")}
}
