% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_spotify_song.R
\name{match_spotify_song}
\alias{match_spotify_song}
\title{Matches Spotify song to corresponding YouTube music video}
\usage{
match_spotify_song(url)
}
\arguments{
\item{url}{Link to Spotify song, including `"https://"` at beginning.}
}
\value{
This function returns a list containing the Spotify popularity score,
 corresponding YouTube video view count, and match confidence score (i.e. the
  confidence that the correct YouTube video was matched to the given Spotify
   song).
}
\description{
Matches Spotify song to corresponding YouTube music video
}
